Clean-up when root directory does not exist
When a root directory does not exist anymore, the artist list is not
cleaned-up.

Since the directory does not exist, we don't try to clean its tags. Note
that it does not impact the tracks and the albums since they are
directly related to a folder thanks to the field `folder_id`.

We directly call the `unlink` method instead of a manual SQL query.