Improve image caching
The Kanban view only contains the computed field. The compute method
takes care of getting the image cache if it exists. By doing so, we take
advantage of the parallelization of the requests, while avoiding the
extra server load due to the real-time computation.

A cron (once a day) is defined in order to keep this cache up-to-date.