Improve image caching
Image caching of small images (used mainly in Kanban view) is simplified
by using a cache system similar to what is done for the LastFM data.
This allows to build the cache 'on-the-fly'.

The button 'Build Image Cache' is therefore not necessary anymore.

Many small errors are also corrected: the computed fields of the images
were working fine for a singleton, but in others cases they were buggy.