Avoid unnecessary reload of view
Make sure to properly remove the bus event listener.

Case 1: list view embedded in form view

Only listen once to the event. It means that the list view will be
updated when the user clicks on it, but the following fired events won't
update the view.

Case 2: 'Current Playlist' view

An issue occurs:
- Go to 'Current Playlist'.
- Click on a line in order to start playing => `reloadFromPlay` is
  called once.
- Go to any other view.
- Go back to 'Current Playlist'.
- Click on a line => `reloadFromPlay` is called twice

When the user switches from one view to the other, the view are not
destroyed (the `destroy` method is not called). Therefore, the web
client is still listening to events which are linked to views which are
not displayed to the user anymore.