Zero value for `maxBitRate`
According to Subsonic API, a zero value for `maxBitRate` implies that no
limit is imposed. Before this commit, the default behavior would not
follow this logic. If a zero value is set, ultimately the transcoder
will use its default bitrate value.

Since the default bitrate value for MP3 conversion is 320 kbps, it means
that a zero value would convert any MP3 file to a 320 kbps, even if the
original bitrate is smaller. This occurs when using the Dsub Android
client, when the bitrate couldn't be determined.

To avoid this issue, we send the original file if no conversion is
needed and the `maxBitRate` is set to zero.