Add loudness normalization option
Normalize following the EBU R128 algorithm, with the following
parameters:
- Programme Loudness:      -18.0 LUFS
- Loudness Range (LRA):      7.0 LU
- Maximum True Peak Level:  -2.0 dBTP

The normalization is made available on Playlists and on Converters.

It should be noted that the normalization is resource-consuming,
therefore it makes the transcoding significantly slower.

This feature requires FFmpeg >=3.2.1 which includes by default the
appropriate library (libebur128).

More info at https://tech.ebu.ch/docs/r/r128.pdf