Stream raw files
Add the possibility to stream raw files, i.e. not transcoded. Since
modern browsers support natively many audio formats, this can be used
without issue for many files.

Raw format is not a default option: it should be activated manually on
the playlist. When activated, the raw format has the highest priority
over all other formats. Howler should be smart enough to select this
format only if it is supported by the browser. Indeed, the file
extension is kept.

In case the track is played from the beginning (`seek=0`), FFmpeg is
bypassed and the track is sent as a regular file thanks to `send_file`.
If the track is seeked, FFmpeg is used with `-c:a copy` option.