Avoid useless write in cache
When the user is playing a track and the folder scan is launched during
the last 30 seconds of the track, the following error can occur during
the final `write` of the scan:
> TransactionRollbackError: could not serialize access due to concurrent
update

When calling `oomusic_next`, we access `image_medium` of the album, and
therefore the related field of the folder. The API magic will also try
prefetching the `image_medium` of the root folder. However, in most
cases this folder doesn't contain a cover and therefore is equal to
`False`. In this case, `_compute_image_medium` will search for a new
image of the folder. It won't find anything, but will still write
`False` on the field `image_medium_cache`.

There comes the issue: it regularly happens that the final `write` of
the folder scan occurs at the same time than the `write` on
`image_medium_cache`. This causes the concurrent update error, and
ultimately leads to a locked folder.

The fix is to avoid any unnecessary write in the image cache
computation.