Add Download and Share feature
Possibitity to download:
- Tracks
- Albums
- Artists
- Genres
- Playlists

In the backend, a 'Download' button is available on the mentioned
objects. When clicked, a ZIP archive containing all related tracks is
created in the temporary folder.

On these models, a 'Download Links' tab is introduced in allow sharing
with external users. A unique token is used in order to grant access to
the tracks. An expiry date is defined on the links; by default, a link
is valid 30 days.

To reduce the server load:
- A minimum delay between consecutive accesses is defined (default: 60
  seconds).
- The name of the ZIP file is deterministic and based on the tracks
  included. A ZIP file is reused when possible.