Add option to deactivate HTML5
Howlerjs provides the Web API and the HTML5 support. HTML5 is better in
our case since it allows playing the song without having to wait for the
complete file. However, this might lead to download errors and premature
ending of the track. We define a field on the playlist to select both
solutions.