Extract user-specific data from music data
Up to now, user-specific data such as the rating, favorite tracks / albums /
artists or the last play date was mixed with the music data (e.g. ID3 tags).

This commit aims to move the user-specific information to a dedicated table.
This will allow sharing the music library among multiple users.