Allow sharing the music library among users
A 'Folder Sharing' option is added to the settings in order to make the tracks
shared among all users. When the option is active, the music folders are
accessible to all users. The sharing possibility works on a all or nothing
logic: either the user can only see the content of his own folders, or he can
see the content of all folders.

In case the option is deactivated after being activated, all playlist lines
and user-specific data (rating, favorites) pointing to tracks owned by another
user will be deleted.