Explicitly unlink tracks and albums
When unlinking a folder, tracks and albums are automatically deleted
thanks to the SQL `CASCADE` mechanism. However, this mechanism bypasses
the usual `unlink` methods of the mentioned models.

This becomes an issue with the introduction of a specific preference
table. Indeed, the preferences are not cleaned up when a folder is
removed, which leads to attempts at retrieving non-existing records.

Instead of relying on the `CASCADE` mechanism, we explicitly unlink the
tracks and albums included.