Optimize computation of `in_playlist`
Every time `in_playlist` is written on a track, we recompute the
corresponding album field thanks to `_inverse_in_playlist`.

However, when adding several tracks to a playlist, each playlist line
creation triggers a `write` on the track. This means that if there are X
tracks on an album, we execute `_inverse_in_playlist` X times.

We do not store `in_playlist` anymore, but on the other hand we optimize
its computation. We only write on the tracks/albums that are in the
current playlist, which reduces the process load. Moreover, the
deactivation of the prefetching on the playlist reduces significantly
the amount of unnecessary data.

The interesting side effect is the removal of several crappy overrides
which were necessary for the sole purpose of the `in_playlist` field.