Migration to v12: set preference
In a shared folder configuration, setting a field defined in
`oomusic.preference` with a user different from the owner of the object
raises an AccessError.

It seems that a `check_access_rule('write')` is performed even if the
user doesn't explicitly write on the object. Therefore, we need to write
as superuser.