Delete folder
When a non-superuser tries to delete a folder (even his own), an
AccessError is raised on `oomusic.tracks`.

A regular user doesn't have the right to delete tracks and albums. This
is to avoid deletion from the interface. However, in this specific case,
it should be allowed to do it.