Automatically unlock folders
It might happen that a folder is locked for no good reason. This can
happen if the server was restarted, or the scanning crashed for some
reason.

We introduce a cron which checks if a locked folder should still be
locked. Since a commit is supposed to be done (in the worst case) every
two minutes, we can determine which folders are in a faulty state.