Add upcoming events of artists
Use the BandsInTown data to display the upcoming events of an artist.

A user can follow an artist using a switch at the top right of the
artist form view. Once an artists is followed, the events are retrieved
automatically thanks to a scheduled action. It is also possible to
manually retrieve the upcoming events, by clicking on the 'Update
Events' button located in the 'Events' tab of the artist form view.

The events are only retrieved for followed artists and automatically
cleaned up.