Deactivate create at view level
There is no need to manually create a suggestion, a track, an album, an
artist or a genre. Therefore, we deactivate at the view level.

For regular users, that doesn't change anything since the ACL does not
allow it, but it is still possible as supersuer.

In the case of edit, it is still necessary to modify a download link, so
we keep the button.