Filter events based on distance
Add the following fields on the user: latitude, longitude and maximum
distance. They are used to filter out events based on their distance
from the user location.

If the maximum distance is not specified, no filtering is performed.
However, it is still possible to add a distance filter in the search
view.