Stop using new cursors
New cursors were introduced for intermediary commits. They are actually
not necessary at all: `self.env.cr.commit()` is enough to commit.

The code simplification is non-negligible, and moreover it avoids doing
weird stuff in the tests.

New cursors are used in a specific case: in the methods called in a
`Thread`.