Direct use of cached image
The listing of folders and albums through the `getAlbumList` or
`getAlbumList2` Subsonic endpoints is quite slow because of the
computation of `image_medium`.

Indeed, `make_Child_folder` and `make_AlbumID3` trigger the method
`_compute_image_medium` for all folders and albums, which is more
time-consuming than expected, even if the image cache already exists.

Therefore, we use the cache image as a first option, and fall back on
the computed file.