Track number
At the moment, sorting by track number is correct in the following
cases:
- 01, 02, 03...
- 1, 2, 3...
- 01/11, 02/11, 03/11...

In the two first cases, an integer value is correctly extracted. In the
third case, the integer is not extracted, but since we also sort on the
string version of the track number, it works.

However, when tracks are numbered as: 1/11, 2/11, 3/11... sorting fails
since the order will be 1/11, 10/11, 12/11, 2/11, 3/11...

We apply the same logic as in `make_Child_track`.

Fixes https://github.com/DocMarty84/koozic/issues/18