Support `ir.actions.act_play`
A new `ir.actions.act_play` is supported by the Action Manager. It's purpose
is to replace the existing Play Field, as well as be used in action button.
More specifically, any button call can now trigger a track playback.

It triggers a `oomusic_play` event in order to play a track or a playlist
track.