Use of `ir.action.act_play`
Replace the PlayField by a button call to an action returning an action
of type `ir.action.act_play`. Moreover, add several action buttons using
the latter in order to start playing tracks from most of the views.

The logic is always 'Add And Play' for albums and artists: add to playlist
and start playing, so the user doesn't clear the current playlist
unintentionally. In case of tracks, we keep the behavior of the PlayField,
i.e. the track is played, but not added to the playlist.

Fixes https://github.com/DocMarty84/koozic/issues/14