Reduce memory usage during folder scanning
When scanning a folder, a lot of DB data is copied to the cache to
reduce the DB access bottleneck. This doesn't make the folder scanning
scalable.

We adopt a more lazy approach: we only fetch the track-related data if
something has changed in the folder. Indeed, if nothing has changed at
the folder level, we don't go further and don't scan the tracks. This is
expected to be the most common use case, since tracks are not likely to
be modified.