`_compute_in_playlist` performance
The most recent performance improvement on `_compute_in_playlist`
assumed that the quantity of albums in the current playlist was smaller
than the quantity of albums displayed.

In case of a dynamic playlist, this might not be the case since tracks
can be added to the playlist regardless of their album. As a
consequence, the computation quickly becomes slower as the size of the
current playlist grows. Indeed, adding the N tracks of an album to the
playlist only adds one album to fetch. However, N tracks added
dynamically are likely to add around N albums to fetch.

We get rid of the `mapped` call and replace it with a single SQL query.
The execution time of the latter as a function of the playlist size is
expected to be constant, as opposed to the current implementation.

Hopefully it's going to be the good solution this time!