Add remote control support
This commit makes possible the playback control thanks to a remote page.
The page is protected by a token and is either public (no login
required) or private (login required).

All controls are available. Setting a track as favorite is only allowed
in private mode.

A QR code containing the complete URL is provided for convenience.