Selection of Subsonic API format
The default format for the Subsonic API is MP3 since it is the only format
supported by Subsonic. However, most modern mobile applications are able
to handle other formats such as OGG or OPUS.

We add the possibility to choose the transcoding format in the settings.

Fixes https://github.com/DocMarty84/koozic/issues/26