Interrupted folder scanning
When the first folder scan is interrupted then restarted,
`cache["track"].keys()` crashes because `track` is not in the keys.

When restarting in such a situation, the first time `skip` is `False` is
on a new folder. But at the same time, `build_cache_folder` is also
`False`. Therefore, `track` is not added in the `cache`, causing the
crash.

We simply make sure that the key is always in the `cache`.