Implement Play/Skip ratio
Up to now, a track was considered played as soon as it was loaded. This
was incorrect in many situations. For example, clicking on 'Stop'
increments the play count.

This commit introduces the following behavior:
- a track is considered played if playback >= 50 % of the track duration
- a track is considered skipped if 'Next' or 'Previous' is clicked and
  playback is < 50 % of the track duration

The Play/Skip ratio of a track can only be updated once every 5 minutes
to account for users coming back and forth on a track.

Closes https://github.com/DocMarty84/koozic/issues/41