Do not force commit in computed fields
The following commit leads to `CacheMissed` errors when retrieving an
album image for the first time:
https://github.com/odoo/odoo/commit/0f6875a5410f18e8f3f639721e774c463090642d

It happens because a call to `commit` also clears the cache.

In order to fix the issue, we could assign the value to the computed
field after the `commit` call. However, the programmatic call to
`commit` is not necessary since it happens automatically at the end of
the transaction.