Prevent unnecessary cache invalidation
We use 2 scheduled actions with a high frequency:
- check conversion queue (every 2 minutes)
- unlock folders (every 10 minutes)

The issue is that the job processing triggers a cache invalidation which
impacts performances.

The conversion queue is now dynamically activated when we run a
converter. It is automatically deactivated once a day if no converter is
running.

The folder unlocking is now executed every 3 hours. Indeed, the 'Unlock'
button is now made more visible, so it can be done maunally. We still
keep the cron to avoid the library not being scanned for ages.